Add files via upload
This is a word scramble game. When the game is played, the computer selects a random word from a text file of possible words. The word is scrambled and printed to the console, and the user attempts to guess the unscrambled word. The user is given one fewer guesses than there are letters in the word. For every incorrect guess, part of the unscrambled word is shown. If the first guess is wrong, the first letter of the word is shown. If the second guess is wrong, the first and second letters are shown, and so on. The user earns 1 point for each letter not shown when the word is guessed correctly. No point is awarded for the last letter, however, since it can be determined by the process of elimination.